processor: single file events workflow
