+++ service/processor/processor_lib/__init__.py
@@ -1,0 +1,5 @@
+from .event_abstraction import EventProcessor
+
+__all__ = [
+    "EventProcessor",
+]
